# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DATN E-Commerce API",
        "description": "Tài liệu API hệ thống DATN E-Commerce Backend",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "DATN API Server"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Giỏ hàng (Cart)"
                ],
                "summary": "Xem sản phẩm trong giỏ hàng",
                "operationId": "0652f94dde92e36bbb3c461356ce1d7b",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Giỏ hàng (Cart)"
                ],
                "summary": "Thêm sản phẩm vào giỏ hàng",
                "operationId": "b432c7bbb8b06ea2019c3d37f23258ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "variant_id"
                                ],
                                "properties": {
                                    "variant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đã thêm vào giỏ hàng"
                    },
                    "400": {
                        "description": "Vượt quá tồn kho"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/{id}": {
            "put": {
                "tags": [
                    "Giỏ hàng (Cart)"
                ],
                "summary": "Cập nhật số lượng sản phẩm trong giỏ hàng",
                "operationId": "ce6d8910efe0dda5a56c85b4383c4010",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cart ID hoặc Variant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Giỏ hàng (Cart)"
                ],
                "summary": "Xóa mục khỏi giỏ hàng",
                "operationId": "5fc80ceb0ebae6f9fed7fc27e0f93592",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cart ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Giỏ hàng (Cart)"
                ],
                "summary": "Xóa toàn bộ giỏ hàng",
                "operationId": "c2cf94402630acad1cc6d317ec806693",
                "responses": {
                    "200": {
                        "description": "Đã xóa toàn bộ giỏ hàng"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập token của bạn dạng: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}